backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
awscli==1.30.1
blinker==1.7.0
boto==2.49.0
boto3==1.29.1
botocore==1.32.1
certifi==2023.7.22
channels==4.0.0
click==8.1.7
colorama==0.4.4
Django==4.2.7
django-cors-headers==4.3.1
django-dynamodb-sessions==0.6
djangorestframework==3.14.0
dnspython==2.4.2
docutils==0.16
dynamodb==1.0.0
dynamorm==0.11.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
marshmallow==3.20.1
orjson==3.9.10
packaging==23.2
promise==2.3
pyasn1==0.5.0
pynamodb==5.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
